# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/comments": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new comment on a post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Create a new comment",
                "parameters": [
                    {
                        "description": "Comment to create",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_dto.CreateCommentDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.SuccessMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/comments/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update an existing comment by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Update a comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated comment data",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_dto.UpdateCommentDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.SuccessMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete an existing comment by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Delete a comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.SuccessMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/posts/{id}/comments": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve all comments associated with a specific post by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Get comments for a specific post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.SuccessResultResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_dto.CommentDetailResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_util.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_fatihesergg_go_social_internal_dto.CommentDetailResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_followed": {
                    "type": "boolean"
                },
                "is_liked": {
                    "type": "boolean"
                },
                "replies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_dto.ReplyResponse"
                    }
                },
                "total_likes": {
                    "type": "integer"
                },
                "total_reply": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_model.User"
                }
            }
        },
...
        "github_com_fatihesergg_go_social_internal_dto.CreateCommentDTO": {
            "type": "object",
            "required": [
                "content",
                "post_id"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 200
                },
                "image": {
                    "type": "string"
                },
                "post_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_fatihesergg_go_social_internal_dto.ReplyResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_followed": {
                    "type": "boolean"
                },
                "is_liked": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "total_like": {
                    "type": "integer"
                },
                "total_reply": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/github_com_fatihesergg_go_social_internal_model.User"
                }
            }
        },
...
        "github_com_fatihesergg_go_social_internal_dto.UpdateCommentDTO": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 200
                },
                "image": {
                    "type": "string"
                }
            }
        },
...
        "github_com_fatihesergg_go_social_internal_model.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_fatihesergg_go_social_internal_util.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {}
            }
        },
        "github_com_fatihesergg_go_social_internal_util.SuccessMessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_fatihesergg_go_social_internal_util.SuccessResultResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {}
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "This is a simple social media API built with Go and Gin.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}